# docs/partner-gate-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: CrewCheck Partner Gate API
  version: 1.1.0
  description: |
    API B2B para consulta de portão e webhooks assinados de mudança de portão.
    A exportação operacional depende de habilitação explícita e de política de redistribuição por classe de fonte.
servers:
  - url: https://crewcheck.online
security:
  - partnerApiKey: []
paths:
...
  /api/admin/partner-api/keys:
    get:
      summary: Listar credenciais de parceiros
      operationId: adminListPartnerKeys
      security:
        - crewCheckAdminJwt: []
      responses:
        '200': { description: Lista de credenciais sem segredos }
    post:
      summary: Emitir credencial de parceiro
      operationId: adminCreatePartnerKey
      security:
        - crewCheckAdminJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, scopes]
              properties:
                partnerEmail: { type: string, format: email }
                label: { type: string, maxLength: 120 }
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [gates:read, webhooks:manage, flights:watch, rosters:write]
      responses:
        '201': { description: API key retornada uma única vez }
...
components:
  securitySchemes:
    partnerApiKey:
      type: http
      scheme: bearer
      bearerFormat: ck_live_...
      description: API key B2B emitida pelo CrewCheck.
    crewCheckAdminJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...